django_rest_framework/django_rest/src/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
Django==4.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.9
hvac==0.11.2
idna==3.3
loguru==0.6.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.2.1
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
validators==0.20.0
